labs/phase1/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gpu_procurement_system"
version = "0.1.0"
description = "Vertex AI L400 Lab 2: Multi-Agent GPU Procurement System"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "pandas",
    "pydantic",
    "pytest",
    # Add your ADK specific dependency here if it's a public package,
    # otherwise it might need to be installed separately.
    "google-cloud-bigquery>=3.10.0",
    "google-cloud-storage>=2.14.0",
    "google-adk",
    "python-dotenv",
    "google-cloud-secret-manager",
    "google-auth",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff"
]

# This section is critical for the 'src' layout
[tool.setuptools.packages.find]
where = ["src"]  # Look for packages inside src
include = ["*"]  # Include all packages found
namespaces = false # Set to true if you plan to use implicit namespace packages

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = [
    "tests",
]